Address code review: fix negated-unsigned, explicit 32-bit wrap, usize cast, add rfind doc
Agent-Logs-Url: https://github.com/microsoft/windows-rs/sessions/d565f191-ddab-4a91-bcf3-9fb969876fcd

+++ crates/libs/rdl/src/clang/cx.rs
@@ -134,9 +134,9 @@ impl TranslationUnit {
                 return vec![];
             }
 
-            let result = (0..n_tokens)
+            let result = (0..n_tokens as usize)
                 .map(|i| {
-                    let token = *tokens.add(i as usize);
+                    let token = *tokens.add(i);
                     let kind = clang_getTokenKind(token);
                     let spelling = to_string(clang_getTokenSpelling(self.0, token));
                     (kind, spelling)
